% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-latex2fun_safe.R
\name{latex2fun_safe}
\alias{latex2fun_safe}
\title{latex2fun_safe}
\usage{
latex2fun_safe(code)
}
\arguments{
\item{code}{: character string in Latex language}
}
\value{
Latex expression translated into R code
}
\description{
latex2fun_safe
}
