% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-latex2fun_safe.R
\name{latex2fun_safe}
\alias{latex2fun_safe}
\title{latex2fun_safe}
\usage{
latex2fun_safe(latex_string)
}
\arguments{
\item{latex_string}{: character string in Latex language}
}
\value{
fun_out : function object built from Latex string input
}
\description{
Works similar to latex2r::latex2fun with additional input constraints
}
